review: rebrand, comment and move to tracing-utils
